group_exercise2/requirements.txt
numpy==1.26.4
matplotlib==3.8.4
svg.path==6.3
opencv-python==4.9.0.80
tslearn==0.6.3